// public/reference-studio/local-src/src/types.ts
export interface ChartConfig {
  id?: string;
  title: string;
  subtitle?: string;
  type: string; // LineChart, BarChart, etc.
  provider: 'google' | 'recharts';
  xAxisKey: string;
  dataKeys: string[];
  zAxisKey?: string;
  options?: any;
  data?: () => any[]; // For dynamic Google Charts data
  videoCount?: number;
  sortType?: 'recent' | 'highest_rated' | 'alphabetical';
  durationType?: 'shorts' | 'long' | 'combined';
}